+++ CHANGELOG.md
@@ -670,6 +670,7 @@
 ## [unreleased]
 
 - add spacing between view and edit links on Organisation index page
+- export complete transaction history per delivery partner for BEIS users
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-52...HEAD
 [release-52]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-51...release-52